included test fixtures and redifined test folder
