+++ packages/multiple-did-identities/src/tests/base64urlencode.spec.ts
