+++ packages/multiple-did-identities/src/tests/convert-to-abbreviated.spec.ts
