+++ packages/multiple-did-identities/src/tests/did-key-method.spec.ts
