+++ packages/multiple-did-identities/src/tests/did-methodfactory.spec.ts
