+++ packages/multiple-did-identities/src/tests/did-peer-3.spec.ts
