+++ packages/multiple-did-identities/src/tests/did-peer-4.spec.ts
