+++ packages/multiple-did-identities/src/tests/did-peer-methodtype-.spec.ts
